Rename integration tests package to avoid confusion re what it tests
